Fix Color

+++ Sudoku/src/cz/krejcar25/sudoku/style/Color.java
@@ -3,31 +3,31 @@ package cz.krejcar25.sudoku.style;
 import processing.core.PGraphics;
 
 public class Color {
-    public byte r;
-    public byte g;
-    public byte b;
-    public byte a;
+    public int r;
+    public int g;
+    public int b;
+    public int a;
 
-    public Color(byte gray) {
+    public Color(int gray) {
         this.r = gray;
         this.g = gray;
         this.b = gray;
     }
 
-    public Color(byte gray, byte a) {
+    public Color(int gray, int a) {
         this.r = gray;
         this.g = gray;
         this.b = gray;
         this.a = a;
     }
 
-    public Color(byte r, byte g, byte b) {
+    public Color(int r, int g, int b) {
         this.r = r;
         this.g = g;
         this.b = b;
     }
 
-    public Color(byte r, byte g, byte b, byte a) {
+    public Color(int r, int g, int b, int a) {
         this.r = r;
         this.g = g;
         this.b = b;
